% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmslib.R
\name{import_internal}
\alias{import_internal}
\title{import internal functions from the KernelKnn package}
\usage{
import_internal(function_name)
}
\description{
import internal functions from the KernelKnn package
}
\keyword{internal}
