% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmslib.R
\name{inner_kernel_function}
\alias{inner_kernel_function}
\title{inner function to compute kernels, extract weights and return predictions}
\usage{
inner_kernel_function(y_matrix, dist_matrix, Levels, weights_function, h)
}
\description{
inner function to compute kernels, extract weights and return predictions
}
\keyword{internal}
